+++ package.json
@@ -93,11 +93,14 @@
     "@babel/preset-typescript": "^7.12.1",
     "@elastic/elasticsearch": "^7.3.0",
     "@types/aws-lambda": "^8.10.64",
+    "@types/debug": "^4.1.5",
     "@types/dockerode": "^2.5.20",
     "@types/fs-extra": "^8.1.0",
     "@types/koa": "^2.0.45",
     "@types/koa-router": "^7.0.28",
     "@types/lodash": "^4.14.150",
+    "@types/nested-error-stacks": "^2.1.0",
+    "@types/promise-retry": "^1.1.3",
     "@types/tmp": "^0.2.0",
     "@types/unzipper": "^0.10.3",
     "@types/uuid": "^7.0.3",
