--- src/utils/config.d.ts
@@ -1,3 +1,0 @@
-import PQueue from "p-queue";
-
-export const pQueue: PQueue;
