+++ src/utils/config.ts
@@ -1,11 +1,7 @@
-const { default: PQueue } = require('p-queue');
+import PQueue from 'p-queue';
 
 const concurrency = process.env.LAMBDA_TOOLS_CONCURRENCY
   ? Number.parseInt(process.env.LAMBDA_TOOLS_CONCURRENCY, 10)
   : Number.POSITIVE_INFINITY;
 
-const pQueue = new PQueue({ concurrency });
-
-module.exports = {
-  get pQueue () { return pQueue; }
-};
+export const pQueue = new PQueue({ concurrency });
