--- src/utils/logging.d.ts
@@ -1,9 +1,0 @@
-export interface Logger {
-  info(...any: any): void;
-  error(...any: any): void;
-  warn(...any: any): void;
-  debug(...any: any): void;
-  child(name: string): Logger;
-}
-
-export function getLogger(name: string): Logger;
