+++ src/utils/logging.ts
@@ -1,24 +1,39 @@
-const createDebug = require('debug');
+import createDebug, { Debugger } from 'debug';
 
 const libName = 'lambda-tools';
 
-const loggers = {};
+const loggers: Record<string, Logger> = {};
 const consoleLog = console.log.bind(console);
 
-function extendLogger ({ logger, name, enabled, log }) {
+export interface Logger {
+  info(...any: any): void;
+  error(...any: any): void;
+  warn(...any: any): void;
+  debug(...any: any): void;
+  child(name: string): Logger;
+}
+
+interface LoggerExtension {
+  logger: Debugger;
+  name: string;
+  enabled?: boolean;
+  log?: typeof console.log;
+}
+
+function extendLogger ({ logger, name, enabled, log }: LoggerExtension) {
   const childLogger = logger.extend(name);
   childLogger.log = log || logger.log;
   childLogger.enabled = enabled || logger.enabled;
   return childLogger;
 }
 
-function createChildLogger (name, root) {
+function createChildLogger (name: string, root: Debugger) {
   return {
     info: extendLogger({ logger: root, name: 'info', log: consoleLog, enabled: true }),
     error: extendLogger({ logger: root, name: 'error', enabled: true }),
     warn: extendLogger({ logger: root, name: 'warn', enabled: true }),
     debug: extendLogger({ logger: root, name: 'debug', log: consoleLog, enabled: root.enabled }),
-    child: (name) => {
+    child: (name: string) => {
       const child = root.extend(name);
       child.enabled = root.enabled;
       return createChildLogger(name, child);
@@ -26,14 +41,10 @@ function createChildLogger (name, root) {
   };
 }
 
-function getLogger (name) {
+export function getLogger (name: string): Logger {
   const fullName = `${libName}:${name}`;
   if (!loggers[fullName]) {
     loggers[fullName] = createChildLogger(fullName, createDebug(fullName));
   }
   return loggers[fullName];
 }
-
-module.exports = {
-  getLogger
-};
